feat - 날짜 직접입력 필터 기능 추가, 코드 분리
